==> 0_Castellario_gender_theme.R <==
################################################################################
## INQUISITION IN GIAVENO (1335)
## (0) Castellario: Gender study (theme for plots)
## R script by Davor Salihovic (Uni of Antwerp) & Jose Luis Estevez (Uni of Helsinki)
## Date: Oct 5th, 2024
################################################################################

# Set theme for plots
theme_paper <- theme_classic() +
            theme(text = element_text(family = "serif"),
                  axis.title.x = element_text(family = "serif"),
                  axis.title.y = element_text(family = "serif"),
                  axis.text.x = element_text(family = "serif"),
                  axis.text.y = element_text(family = "serif"),
                  legend.text = element_text(family = "serif"),
                  plot.subtitle = element_text(hjust = 0.5),
                  plot.title = element_text(family = "serif"))

################################################################################
